Streamlining the graphics interface more and more.
Added the VulkanViewport class which will store the swapchain. Currently it just acts as a holder, but it should eventually be responsible for more than just the swapchain.

Removed the Renderer class as well. It was unnecessary in its current form. It might come back later on as some sort of middle ground for interfacing with the graphics api