Beginning to move the vulkan behavior out of the high level render pipeline and down to the lower level vulkan interfaces.
There is now a separate graphics interface call for uploading texture data to a native texture and transitioning any texture to write or read states. There will soon be more, I'm sure

Added a higher level Viewport class as well. Might be the same sort of function as the ScreenView class later on, but we will see