Fixed some bugs in ResourceBlob, now have to Initialize and deinitialize shader compiler.
Updated local vulkan sdk because the version 1.0.39.1 was giving me trouble on my desktop for some reason (more specifically, calls to vkGetSwapchainImagesKHR were resulting in incorrect handles to the swapchain images. This started a few weeks ago potentially, but I don't know why this could be happening. Probably driver issue)

Updated to 1.1.85.0 on laptop and desktop but not actually using that SDK yet in Musa. Fixed issues that came along with it