Starting to implement a logging implementation. This will live in Core, becasuse we want all parts of the engine to have log capabilities.
This addition also means that I needed to implement some threading behavior, that will surely need to be looked at later on, and the elimination of the File dll, which moved the File IO behavior into Core. Logging needs to have some sort of file IO, so therefore, it needs to be in Core. It being in Core will allow more platform specific stuff to not be in multiple places in the engine. Makes more sense going forward