All of the output sinks are now working fine. For the logging loop, I am now using a semaphore to signal when there is something in the queue that needs processing.
Logging has been introduced to the graphics callback for Vulkan. It now logs everything to come through vulkan. Currently, it's just debugging information, but it is working very well! Now to introduce it to a wider portion of the engine (and also have a way to shut down logging to capture all of the information!)