Getting back to ECS integration.
More importantly, fixed a memory allocation issue with realloc. It was using the original size of the block of memory instead of the new size of the realloc request. This led to memory corruption when asking Realloc for a smaller piece of memory than was actually there, so it memcpyed a bigger chunk of memory to a location that was supposed to have a smaller chunk of memory