Huuuuuuuge changes. Lots of little and bit changes all around.
Main goal was to have a way to push windows events into the ECS ecosystem. Lots of changes to get that to work, even though it's not even done yet. Want to push this up just so I have a jumping off point for getting it to actually work.

One input is set up which is the escape key closing the program.