Working on getting input officially working for parts of the ECS ecosystem. Main thing right now is just getting the GodCamera working.
One thing that needs to be focused on is geting events OSEvents/ApplicationEvents back to the application part of the engine. In the system outside of the ECS ecosystem, it worked when an input came up because an InputEvent was given back to the dispatcher, which then was processed. Can't process that kind of stuff immediately like that, might want to have a system at the end of the frame to process existing events like that?

I think that covers all of the issues coming up, bound to be more later on with this kind of integration