Completed the render integration of ImGui, at least to my knowledge
Drawing via the graphics interface has changed, due to the need for using a specific amount of an index buffer instead of the entire buffer. The api requires an amount of primitives, graphics primitives, to be passed in. Right now, this only supports triangle list, which is an issue. This will just lead to a general refactor of the graphics api in general, but I would like some sort of editor first.

Hopefully all Imgui stuff doesn't touch rendering anymore, but we will see!